refactor(middleware): restructure middleware handling and security
- Move middleware logic to centralized handler in middleware directory
- Consolidate rate limiting and security headers into middleware composition
- Remove old middleware.ts file in favor of new implementation
