+++ middleware.ts
@@ -1,0 +1,20 @@
+// Import and re-export the middleware from the middleware directory
+import { middleware as middlewareHandler } from '@/middleware';
+import { NextRequest } from 'next/server';
+
+export async function middleware(request: NextRequest) {
+  return middlewareHandler(request);
+}
+
+export const config = {
+  matcher: [
+    /*
+     * Match all request paths except for the ones starting with:
+     * - _next/static (static files)
+     * - _next/image (image optimization files)
+     * - favicon.ico (favicon file)
+     * - public folder
+     */
+    '/((?!_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp)$).*)',
+  ],
+};
